Fix player item merging logic
lodash.merge tries to merge arrays inside of the object, but this can
lead to items being incorrect, since it merges based on the index of the
value in the array. for items, this would mean [1, 2, 3] as the base,
with [1, 3] would turn into [1, 3, 3], since it won't erase the last
property. now i've added logic to outright replace it.

glad i caught this before NA/EU LCS! ty rito for broadcasting two test
games worth of data to debug with :D i'll pretend they did it for me...
;)